Extract varisat-internal-proof crate
