+++ varisat-internal-macros/Cargo.toml
@@ -3,7 +3,7 @@ name = "varisat-internal-macros"
 version = "0.2.0"
 authors = ["Jannis Harder <me@jix.one>"]
 edition = "2018"
-description = "Internal macros for varisat"
+description = "Internal macros for the Varisat SAT solver"
 homepage = "https://jix.one/project/varisat/"
 repository = "https://github.com/jix/varisat"
 license = "MIT/Apache-2.0"
