+++ varisat-internal-proof/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "varisat-internal-proof"
+version = "0.2.0"
+authors = ["Jannis Harder <me@jix.one>"]
+edition = "2018"
+description = "Internal proof format for the Varisat SAT solver"
+homepage = "https://jix.one/project/varisat/"
+repository = "https://github.com/jix/varisat"
+license = "MIT/Apache-2.0"
+readme = "README.md"
+
+[dependencies]
+failure = "0.1.5"
+
+    [dependencies.varisat-formula]
+    path = "../varisat-formula"
+    version = "=0.2.0"
+
+[dev-dependencies]
+proptest = "0.9.3"
