+++ varisat-internal-proof/README.md
@@ -1,0 +1,24 @@
+# Varisat
+
+This crate defines the proof format used internally by the [`varisat`
+crate][crate-varisat].
+
+## License
+
+The Varisat source code is licensed under either of
+
+  * Apache License, Version 2.0
+    ([LICENSE-APACHE](LICENSE-APACHE) or
+    http://www.apache.org/licenses/LICENSE-2.0)
+  * MIT license
+    ([LICENSE-MIT](LICENSE-MIT) or http://opensource.org/licenses/MIT)
+
+at your option.
+
+### Contribution
+
+Unless you explicitly state otherwise, any contribution intentionally submitted
+for inclusion in Varisat by you, as defined in the Apache-2.0 license, shall be
+dual licensed as above, without any additional terms or conditions.
+
+[crate-varisat]: https://crates.io/crates/varisat
