+++ varisat-internal-proof/src/vli_enc.rs
