+++ varisat/Cargo.toml
@@ -35,6 +35,10 @@ vec_mut_scan = "0.1.0"
     path = "../varisat-dimacs"
     version = "=0.2.0"
 
+    [dependencies.varisat-internal-proof]
+    path = "../varisat-internal-proof"
+    version = "=0.2.0"
+
 [dev-dependencies]
 proptest = "0.9.3"
 rand = "0.6.5"
