+++ varisat/src/binary.rs
@@ -3,9 +3,10 @@
 use partial_ref::{partial, PartialRef};
 
 use varisat_formula::Lit;
+use varisat_internal_proof::{DeleteClauseProof, ProofStep};
 
 use crate::context::{AssignmentP, BinaryClausesP, Context, ProofP, SolverStateP};
-use crate::proof::{self, DeleteClauseProof, ProofStep};
+use crate::proof;
 
 /// Binary clauses.
 #[derive(Default)]
