+++ varisat/src/cdcl.rs
@@ -3,6 +3,7 @@
 use partial_ref::{partial, PartialRef};
 
 use varisat_formula::Lit;
+use varisat_internal_proof::ProofStep;
 
 use crate::analyze_conflict::analyze_conflict;
 use crate::clause::{assess_learned_clause, bump_clause, db, decay_clause_activities};
@@ -12,7 +13,7 @@ use crate::context::{
 };
 use crate::decision::make_decision;
 use crate::incremental::{enqueue_assumption, EnqueueAssumption};
-use crate::proof::{self, ProofStep};
+use crate::proof;
 use crate::prop::{backtrack, enqueue_assignment, propagate, Conflict, Reason};
 use crate::simplify::{prove_units, simplify};
 use crate::state::SatState;
