+++ varisat/src/load.rs
@@ -4,13 +4,14 @@ use vec_mut_scan::VecMutScan;
 use partial_ref::{partial, PartialRef};
 
 use varisat_formula::Lit;
+use varisat_internal_proof::{DeleteClauseProof, ProofStep};
 
 use crate::clause::{db, ClauseHeader, Tier};
 use crate::context::{
     AssignmentP, BinaryClausesP, ClauseAllocP, ClauseDbP, Context, ImplGraphP, IncrementalP,
     ProofP, SolverStateP, TmpDataP, TrailP, VsidsP, WatchlistsP,
 };
-use crate::proof::{self, DeleteClauseProof, ProofStep};
+use crate::proof;
 use crate::prop::{assignment, full_restart, Reason};
 use crate::simplify::resurrect_unit;
 use crate::state::SatState;
