+++ varisat/src/simplify.rs
@@ -3,6 +3,7 @@
 use partial_ref::{partial, split_borrow, PartialRef};
 
 use varisat_formula::Lit;
+use varisat_internal_proof::{clause_hash, lit_hash, DeleteClauseProof, ProofStep};
 
 use crate::binary::simplify_binary;
 use crate::clause::db::filter_clauses;
@@ -10,7 +11,7 @@ use crate::context::{
     AssignmentP, BinaryClausesP, ClauseAllocP, ClauseDbP, Context, ImplGraphP, ProofP,
     SolverStateP, TrailP, WatchlistsP,
 };
-use crate::proof::{self, clause_hash, lit_hash, DeleteClauseProof, ProofStep};
+use crate::proof;
 use crate::prop::{enqueue_assignment, Reason};
 
 /// Remove satisfied clauses and false literals.
